clean up tri-complete
